refactor: move B3Propagator to B3MultiPropagator
